Fix PlatformViewWrapperTest
